requirements.txt
parsel==1.10.0
httpx==0.28.1
beautifulsoup4==4.13.4
httpx[http2]==4.2.0
selenium==4.33.0
pytest-playwright==0.7.0